base/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
django-jinja==2.11.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==79.0.1
simplejson==3.20.1
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
